# bachpals-ictsss-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend to the ICTSSS service.",
        "title": "ICT-Stack Self-Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ICTSSS Support",
            "email": "\u003cTBD\u003e"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/api/v1",
    "paths": {
...
        "/image/": {
            "get": {
                "description": "Retrieves a list of images that can be used by admins",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Retrieve list of images",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Images"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "Handles the updates on an image",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Updates an image",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "image",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UpdateImageStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Images"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Adds a new image for use in the service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Add a new image",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "image",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AddImageStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Images"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "Deletes an image from the service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Deletes an image",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ImageIdStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Images"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/image/:id": {
            "get": {
                "description": "Retrieves a specific image",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Retrieves an image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "501": {
                        "description": "Not Implemented"
                    }
                }
            }
        },
        "/image/config": {
            "get": {
                "description": "Fetches config files from hard drive",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Fetches config files",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/image/published": {
            "get": {
                "description": "Fetches images marked as published from DB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Fetches published images",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v1.PublishedImagesStruct"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/image/server": {
            "get": {
                "description": "Fetches images from OpenStack",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Fetches images",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Images"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "database.Images": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "imageConfig": {
                    "type": "string"
                },
                "imageDescription": {
                    "type": "string"
                },
                "imageDisplayName": {
                    "type": "string"
                },
                "imageId": {
                    "type": "string"
                },
                "imageName": {
                    "type": "string"
                },
                "imageReadRootPassword": {
                    "type": "boolean"
                },
                "published": {
                    "type": "string"
                }
            }
        },
...
        "v1.AddImageStruct": {
            "type": "object",
            "properties": {
                "image_config": {
                    "type": "string"
                },
                "image_description": {
                    "type": "string"
                },
                "image_display_name": {
                    "type": "string"
                },
                "image_id": {
                    "type": "string"
                },
                "image_name": {
                    "type": "string"
                },
                "image_read_root_password": {
                    "type": "boolean"
                },
                "published": {
                    "type": "string"
                }
            }
        },
        "v1.ImageIdStruct": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "v1.PublishedImagesStruct": {
            "type": "object",
            "properties": {
                "imageDisplayName": {
                    "type": "string"
                },
                "imageId": {
                    "type": "string"
                }
            }
        },
...
        "v1.UpdateImageStruct": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image_config": {
                    "type": "string"
                },
                "image_description": {
                    "type": "string"
                },
                "image_display_name": {
                    "type": "string"
                },
                "image_id": {
                    "type": "string"
                },
                "image_name": {
                    "type": "string"
                },
                "image_read_root_password": {
                    "type": "boolean"
                },
                "published": {
                    "type": "string"
                }
            }
        }
    }
}